requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
groq==0.11.0
python-multipart==0.0.9
websockets==12.0
orjson==3.10.7
markdown==3.7
pdfminer.six==20240706
PyPDF2==3.0.1
google-generativeai>=0.7.0
